requirements.txt
discord.py[voice]==2.3.2
flask==3.0.0
python-dotenv==1.0.0
firebase-admin==6.3.0
yt-dlp==2024.4.9
PyNaCl==1.5.0
groq==0.4.0
ffmpeg-python==0.2.0